chore: track type coercion via scope-level variable
This is not the nicest but the easiest way, we could also detect this package being used by walking the AST, but I wanted to decouple that particular change from this PR
