+++ packages/@cdktf/hcl2cdk/lib/generation.ts
@@ -35,6 +35,7 @@ import {
   getBlockTypeAtPath,
   getAttributeTypeAtPath,
   getTypeAtPath,
+  getDesiredType,
 } from "./provider";
 import { coerceType } from "./coerceType";
 
@@ -61,6 +62,77 @@ function getReference(graph: DirectedGraph, id: string) {
   }
 }
 
+function destructureAst(ast: t.Expression): string[] | undefined {
+  switch (ast.type) {
+    case "Identifier":
+      return [ast.name];
+    case "MemberExpression":
+      const object = destructureAst(ast.object);
+      const property = destructureAst(ast.property as t.Expression);
+      if (object && property) {
+        return [...object, ...property];
+      } else {
+        return undefined;
+      }
+    default:
+      return undefined;
+  }
+}
+
+function findTypeOfReference(
+  scope: Scope,
+  ast: t.Expression,
+  references: Reference[]
+): AttributeType {
+  const isReferenceWithoutTemplateString =
+    ast.type === "MemberExpression" &&
+    ast.object.type === "Identifier" &&
+    references.length === 1;
+  // TODO: from type is not always string, could be any in case of locals or whatever in case of references
+  // If we only have one reference this is a
+  if (isReferenceWithoutTemplateString) {
+    // TODO: check if locals == variables
+    if (references[0].isVariable) {
+      return "dynamic";
+    }
+
+    const destructuredAst = destructureAst(ast);
+    if (!destructuredAst) {
+      logger.debug(
+        `Could not destructure ast: ${JSON.stringify(ast, null, 2)}`
+      );
+      return "dynamic";
+    }
+
+    const [astVariableName, ...attributes] = destructuredAst;
+    const variable = Object.values(scope.variables).find(
+      (x) => x.variableName === astVariableName
+    );
+
+    if (!variable) {
+      // We don't know, this should not happen, but if it does we assume the worst case and make it dynamic
+      return "dynamic";
+    }
+
+    const { resource: resourceType } = variable;
+    const [provider, ...resourceNameFragments] = resourceType.split("_");
+    const tfResourcePath = `${provider}.${resourceNameFragments.join(
+      "_"
+    )}.${attributes.map((x) => toSnakeCase(x)).join(".")}`;
+    const type = getTypeAtPath(scope.providerSchema, tfResourcePath);
+
+    // If this is an attribute type we can return it
+    if (typeof type === "string" || Array.isArray(type)) {
+      return type;
+    }
+
+    // Either nothing is found or it's a block type
+    return "dynamic";
+  }
+
+  return "string";
+}
+
 export const valueToTs = async (
   scope: Scope,
   item: TerraformResourceBlock,
@@ -79,112 +151,6 @@ export const valueToTs = async (
         scopedIds
       );
       const ast = referencesToAst(scope, item, references, scopedIds);
-
-      // Move const into function
-      function findTypeOfReference(
-        scope: Scope,
-        ast: t.Expression,
-        references: Reference[]
-      ): AttributeType {
-        const isReferenceWithoutTemplateString =
-          ast.type === "MemberExpression" &&
-          ast.object.type === "Identifier" &&
-          references.length === 1;
-        // TODO: from type is not always string, could be any in case of locals or whatever in case of references
-        // If we only have one reference this is a
-        if (isReferenceWithoutTemplateString) {
-          // TODO: check if locals == variables
-          if (references[0].isVariable) {
-            return "dynamic";
-          }
-
-          function destructureAst(ast: t.Expression): string[] | undefined {
-            switch (ast.type) {
-              case "Identifier":
-                return [ast.name];
-              case "MemberExpression":
-                const object = destructureAst(ast.object);
-                const property = destructureAst(ast.property as t.Expression);
-                if (object && property) {
-                  return [...object, ...property];
-                } else {
-                  return undefined;
-                }
-              default:
-                return undefined;
-            }
-          }
-
-          const destructuredAst = destructureAst(ast);
-          if (!destructuredAst) {
-            logger.debug(
-              `Could not destructure ast: ${JSON.stringify(ast, null, 2)}`
-            );
-            return "dynamic";
-          }
-
-          const [astVariableName, ...attributes] = destructuredAst;
-          const variable = Object.values(scope.variables).find(
-            (x) => x.variableName === astVariableName
-          );
-
-          if (!variable) {
-            // We don't know, this should not happen, but if it does we assume the worst case and make it dynamic
-            return "dynamic";
-          }
-
-          const { resource: resourceType } = variable;
-          const [provider, ...resourceNameFragments] = resourceType.split("_");
-          const tfResourcePath = `${provider}.${resourceNameFragments.join(
-            "_"
-          )}.${attributes.map((x) => toSnakeCase(x)).join(".")}`;
-          const type = getTypeAtPath(scope.providerSchema, tfResourcePath);
-
-          // If this is an attribute type we can return it
-          if (typeof type === "string" || Array.isArray(type)) {
-            return type;
-          }
-
-          // Either nothing is found or it's a block type
-          return "dynamic";
-        }
-
-        return "string";
-      }
-
-      function getDesiredType(scope: Scope, path: string): AttributeType {
-        const attributeType = getTypeAtPath(scope.providerSchema, path);
-
-        // Attribute type is not defined
-        if (!attributeType) {
-          return "dynamic";
-        }
-
-        // Primitive attribute type
-        if (typeof attributeType === "string") {
-          return attributeType;
-        }
-
-        // Complex attribute type
-        if (Array.isArray(attributeType)) {
-          return attributeType;
-        }
-
-        // Schema
-        if ("version" in attributeType) {
-          return "dynamic";
-        }
-
-        // Block type
-        console.log(
-          `Found block type for ${path}: ${JSON.stringify(
-            attributeType,
-            null,
-            2
-          )}`
-        );
-        return "dynamic";
-      }
 
       return coerceType(
         scope,
