+++ packages/@cdktf/hcl2cdk/lib/types.ts
@@ -17,4 +17,5 @@ export type Scope = {
       variableName: string;
     }
   >;
+  hasTokenBasedTypeCoercion: boolean;
 };
